extractdata
